requirements.txt
streamlit
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
pydantic>=2.10.0
pydantic-core
python-dotenv
sentence-transformers
numpy
pandas
